configure: silence AC_INIT when building outside a git tree
When autoreconfiguring outside a git tree, there are lots of spurious
git fatal errors because there is no git tree to work on to retrieve the
version information:

    fatal: Not a git repository (or any parent up to mount point /home/ymorin/dev)
    Stopping at filesystem boundary (GIT_DISCOVERY_ACROSS_FILESYSTEM not set).

Or, with an older git version:
    http://autobuild.buildroot.org/results/9df/9df2c09963b97524621f8323b279b8a1672d4b6c/build-end.log

This is rather discouraging for the unsuspecting user, and even seasoned
developpers would have to look twice before they can conclude that's not
an actual error.

Since the version srring is not strictly required for autoconf to work
properly, just consign git's sdterr to oblivion.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>

---
In the Buildroot project, we grab (Aha!) the v4l2grab sources directly
from a tarball generated on the fly by Github, we don't clone the
repository. Thus we have this nasty message, like so: