Use stdint.h types and inttypes.h formats.
This fixes an issue with formatting timestamps on 32-bit platforms,
including the Raspberry Pi, where sizeof(long) = 4.
Likely, the code was developed on a platform with sizeof(long) = 8,
meaning things worked out by chance.

While at it, also calculate the required memory capacity to hold the
filename by using snprintf(), which returns the number of bytes that
would have been written if the passed buffer was big enough.